# src/backend/src/ApplicationManagerAPI/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Application Manager API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Offer/{jobApplicationId}": {
      "get": {
        "tags": [
          "Offer"
        ],
        "operationId": "Offer_GetByJobApplicationId",
        "parameters": [
          {
            "name": "jobApplicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Offer": {
      "post": {
        "tags": [
          "Offer"
        ],
        "operationId": "Offer_Create",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOfferRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Offer/{id}": {
      "put": {
        "tags": [
          "Offer"
        ],
        "operationId": "Offer_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOfferRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Offer"
        ],
        "operationId": "Offer_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "OfferDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "jobApplicationId": {
            "type": "string",
            "format": "guid"
          },
          "status": {
            "$ref": "#/components/schemas/OfferStatus"
          },
          "annualSalary": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "OfferStatus": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Pending",
          "Accepted",
          "Declined"
        ],
        "enum": [
          "Pending",
          "Accepted",
          "Declined"
        ]
      },
...
      "CreateOfferRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "jobApplicationId"
        ],
        "properties": {
          "jobApplicationId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          },
          "status": {
            "$ref": "#/components/schemas/OfferStatus"
          },
          "annualSalary": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpdateOfferRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "jobApplicationId"
        ],
        "properties": {
          "jobApplicationId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          },
          "status": {
            "$ref": "#/components/schemas/OfferStatus"
          },
          "annualSalary": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}